requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.9.1
asteval==1.0.6
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
boto3==1.39.13
botocore==1.39.13
Brotli==1.1.0
bw-migrations==0.2
bw2calc==2.1
bw2data==4.5
bw2io==0.9.11
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
constructive_geometries==1.0
country_converter==1.3.1
cryptography==45.0.5
cssselect2==0.8.0
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.18
dj-database-url==3.0.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-unfold==0.60.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
docopt==0.6.2
dotenv==0.9.9
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.3
idna==3.10
ijson==3.4.0
isodate==0.7.2
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
linear-tsv==1.1.0
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
matrix_utils==0.6
morefs==0.2.2
mrio-common-metadata==0.2.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
ordered-set==4.1.0
pandas==2.3.1
peewee==3.18.2
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
prettytable==3.16.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
pyecospold==4.0.0
PyJWT==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python3-openid==3.2.0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.2
randonneur==0.6.2
randonneur_data==0.6
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==2.0.0
rpds-py==0.26.0
s3transfer==0.13.1
scipy==1.16.0
six==1.17.0
snowflake-id==1.0.2
social-auth-app-django==5.5.1
social-auth-core==4.7.0
SPARQLWrapper==2.0.0
sparse==0.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stats_arrays==0.7
structlog==25.4.0
tableschema==1.21.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unfold==1.2.3
unicodecsv==0.14.1
urllib3==2.5.0
voluptuous==0.15.2
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
wurst==0.4
xlrd==2.0.2
xlsxwriter==3.2.5
zopfli==0.2.3.post1
